thinking I figured out how to lay out module
