+++ python/ry/ryo3/reqwest.pyi
@@ -4,7 +4,8 @@ from http import HTTPStatus
 import ry
 
 if t.TYPE_CHECKING:
-    from ry.ryo3 import URL, Duration, Headers
+    from ry.http import Headers
+    from ry.ryo3 import URL, Duration
 
 class HttpClient:
     def __init__(
